--- src/lab/mcp/client/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from .base_mcp_interface import MCPHandlerInterface
-from .timeemi import TimeemiMCPHandler
-from .vision import VisionMCPHandler
-
-__all__ = ["MCPHandlerInterface", "TimeemiMCPHandler", "VisionMCPHandler"]
