+++ src/lab/mcp/tool_registry.py
@@ -1,0 +1,171 @@
+from __future__ import annotations
+
+import json
+from dataclasses import dataclass
+from typing import TYPE_CHECKING
+
+from lab.mcp._typing import (
+    GetDateAndTimeArgs,
+    GetDateAndTimeResult,
+    RollDiceArgs,
+    RollDiceByTimeArgs,
+    RollDiceByTimeResult,
+    RollDiceResult,
+    ToolTraceItem,
+    UnknownArgs,
+    UnknownResult,
+)
+
+if TYPE_CHECKING:
+    from pydantic import BaseModel
+
+# =============================================================================
+# 4) ToolRegistry：强类型解析入口（你以后扩展就在这里加分支）
+# =============================================================================
+
+
+@dataclass(frozen=True)
+class ParsedTool:
+    """
+    { full_name: timeemi__get_date_and_time,
+      server: timeemi,
+      name: get_date_and_time,
+      args_model: GetDateAndTimeArgs
+    }
+    """
+
+    full_name: str
+    server: str
+    name: str
+    args_model: BaseModel
+
+
+class ToolRegistry:
+    """
+    工具注册表（强类型入口）。
+
+    你要扩展新工具，照着已有分支加：
+    - Args(BaseModel)
+    - Result(BaseModel)
+    - parse_args / parse_result 各加一个 elif
+    """
+
+    @staticmethod
+    def parse_args(full_name: str, arguments_json: str | None) -> ParsedTool:
+        """
+        解析 tool_call.arguments（JSON 字符串）为对应 Args(BaseModel)。
+
+        输入示例：
+            full_name="timeemi__roll_dice"
+            arguments_json='{"n_dice": 3}'
+
+        输出示例：
+            ParsedTool(..., args_model=RollDiceArgs(n_dice=3))
+        """
+        server, name = full_name.split("__", 1)
+        s = arguments_json or "{}"
+
+        if full_name == "timeemi__get_date_and_time":
+            args_model = GetDateAndTimeArgs.model_validate_json(s)
+            return ParsedTool(full_name, server, name, args_model)
+
+        if full_name == "timeemi__roll_dice":
+            args_model = RollDiceArgs.model_validate_json(s)
+            return ParsedTool(full_name, server, name, args_model)
+
+        if full_name == "timeemi__roll_dice_by_current_time":
+            args_model = RollDiceByTimeArgs.model_validate_json(s)
+            return ParsedTool(full_name, server, name, args_model)
+
+        # 未知工具：保底当 dict
+        try:
+            raw = json.loads(s)
+            if not isinstance(raw, dict):
+                raw = {}
+        except Exception:
+            raw = {}
+        args_model = UnknownArgs(data=raw)  # type: ignore
+        return ParsedTool(full_name, server, name, args_model)
+
+    @staticmethod
+    def parse_result(full_name: str, call_tool_result: object) -> BaseModel:
+        """
+        将 FastMCP call_tool 返回转成 Result(BaseModel)。
+
+        说明：
+        - 若 is_error=True，则返回 UnknownResult(data=错误文本)
+        - 否则优先读取 result.data（FastMCP 已反序列化）
+
+        已知工具强校验：
+        - get_date_and_time: data 必须是 str 且符合格式
+        - roll_dice: data 必须是 list[int] 且 1..6
+        """
+        is_error = bool(getattr(call_tool_result, "is_error", False))
+        if is_error:
+            # 尽量提取错误文本
+            blocks = getattr(call_tool_result, "content", None) or []  # type: ignore
+            err_text = None
+            for b in blocks:  # type: ignore
+                t = getattr(b, "text", None)  # type: ignore
+                if t:
+                    err_text = t
+                    break
+            return UnknownResult(data=err_text or "tool_error")
+
+        data = getattr(call_tool_result, "data", None)
+
+        if full_name == "timeemi__get_date_and_time":
+            if not isinstance(data, str):
+                raise TypeError(f"timeemi.get_date_and_time expects str, got {type(data)}")
+            return GetDateAndTimeResult(datetime=data)
+
+        if full_name == "timeemi__roll_dice":
+            if not (isinstance(data, list) and all(isinstance(x, int) for x in data)):  # type: ignore
+                raise TypeError(f"timeemi.roll_dice expects list[int], got {type(data)} {data}")  # type: ignore
+            return RollDiceResult(numbers=data)  # type: ignore
+
+        if full_name == "timeemi__roll_dice_by_current_time":
+            if not (isinstance(data, dict) and "numbers" in data):
+                raise TypeError(
+                    f"timeemi.roll_dice_by_current_time expects dict with 'numbers', got {type(data)} {data}"  # type: ignore
+                )  # type: ignore
+            return RollDiceByTimeResult(**data)  # type: ignore
+
+        return UnknownResult(data=data)
+
+    @staticmethod
+    def tool_content_for_tool_model(full_name: str, result_model: BaseModel) -> str:
+        """
+        生成回填给 Tool Model 的 tool message content（尽量短）。
+
+        注意：你不希望 client 侧做口语化，所以这里返回“原始但简短”的值。
+
+        输出示例：
+            "2026-01-27 20:54:37"
+            "[5, 3, 1]"
+        """
+        if isinstance(result_model, GetDateAndTimeResult):
+            return result_model.datetime
+        if isinstance(result_model, RollDiceResult):
+            return json.dumps(result_model.numbers, ensure_ascii=False)
+        d = result_model.model_dump(exclude_none=True)
+        return json.dumps(d, ensure_ascii=False, default=str)
+
+    @staticmethod
+    def trace_item(parsed: ParsedTool, result_model: BaseModel, *, ok: bool, error: str | None) -> ToolTraceItem:
+        """
+        构造结构化 trace（给 Chat Model）。
+
+        - args：Args(BaseModel) dump 成 dict
+        - raw_result：Result(BaseModel) dump 成 dict
+        """
+        args_dict = parsed.args_model.model_dump(exclude_none=True)
+        raw_dict = result_model.model_dump(exclude_none=True)
+        return ToolTraceItem(
+            server=parsed.server,
+            name=parsed.name,
+            args=args_dict,
+            raw_result=raw_dict,
+            ok=ok,
+            error=error,
+        )
